pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trello-task-maker"
version = "1.0.0"
description = "CLI tool for creating Trello cards from JSON payload"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Vladyslav", email = "vladpanchenko280@gmail.com"}
]
keywords = ["trello", "cli", "task-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.25.0",
    "jsonschema>=3.2.0",
    "python-dotenv>=0.19.0",
]

[project.scripts]
trello-task-maker = "trello_task_maker.main:cli"

[project.urls]
Homepage = "https://github.com/vladyslav-panchenko280/trello-agent-script"
Repository = "https://github.com/vladyslav-panchenko280/trello-agent-script"

[tool.setuptools.packages.find]
where = ["."]
include = ["trello_task_maker*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.md"]
